submitting lesson 1.2
