+++ src/App.js
@@ -1,20 +1,13 @@
 import React from "react";
-
-var todolist = [
-  { id: "1", title: "work" },
-  { id: "2", title: "eat" },
-  { id: "3", title: "sleep" },
-];
+import AddTodoForm from "./AddTodoForm";
+import TodoList from "./TodoList";
 
 function App() {
   return (
     <div>
       <h1>To Do List</h1>
-      <ul>
-        {todolist.map(function (item) {
-          return <li key={item.id.title}>{item.title}</li>;
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
